requirements.txt
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.0
marshmallow==3.0.1
marshmallow-sqlalchemy==0.17.0
numpy==1.17.0
psycopg2==2.8.2
pytest==5.0.1
pytest-flask==0.15.0
python-dateutil==2.8.0
